+++ tests/Stub/UnionTypes.php
@@ -6,7 +6,11 @@ namespace Yiisoft\Proxy\Tests\Stub;
 
 final class UnionTypes
 {
-    public function run(string|int|null $param): void
+    public function param(string|int|null $param): void
+    {
+    }
+
+    public function result(): string|int|null
     {
     }
 }